DL-19223: Migrate repo to scala 3.3.7 minimally

+++ app/repositories/SessionRepository.scala
@@ -45,7 +45,7 @@ class SessionRepository @Inject() (
           Indexes.ascending("lastUpdated"),
           IndexOptions()
             .name("lastUpdatedIdx")
-            .expireAfter(appConfig.cacheTtl, TimeUnit.SECONDS)
+            .expireAfter(appConfig.cacheTtl.toLong, TimeUnit.SECONDS)
         )
       )
     ) {
